com.rest.elevenlabs 1.3.2 (#14)
- Added voice equality operators
- Added better parameter checks in VoicesEndPoint

+++ Runtime/Voices/VoicesEndpoint.cs
@@ -93,6 +93,11 @@ namespace ElevenLabs.Voices
         /// <returns><see cref="VoiceSettings"/>.</returns>
         public async Task<VoiceSettings> GetVoiceSettingsAsync(string voiceId, CancellationToken cancellationToken = default)
         {
+            if (string.IsNullOrWhiteSpace(voiceId))
+            {
+                throw new ArgumentNullException(nameof(voiceId));
+            }
+
             var response = await Api.Client.GetAsync(GetUrl($"/{voiceId}/settings"), cancellationToken);
             var responseAsString = await response.ReadAsStringAsync();
             return JsonConvert.DeserializeObject<VoiceSettings>(responseAsString, Api.JsonSerializationOptions);
@@ -107,6 +112,11 @@ namespace ElevenLabs.Voices
         /// <returns><see cref="Voice"/>.</returns>
         public async Task<Voice> GetVoiceAsync(string voiceId, bool withSettings = true, CancellationToken cancellationToken = default)
         {
+            if (string.IsNullOrWhiteSpace(voiceId))
+            {
+                throw new ArgumentNullException(nameof(voiceId));
+            }
+
             var response = await Api.Client.GetAsync(GetUrl($"/{voiceId}?with_settings={withSettings}"), cancellationToken);
             var responseAsString = await response.ReadAsStringAsync();
             return JsonConvert.DeserializeObject<Voice>(responseAsString, Api.JsonSerializationOptions);
@@ -121,6 +131,11 @@ namespace ElevenLabs.Voices
         /// <returns>True, if voice settings was successfully edited.</returns>
         public async Task<bool> EditVoiceSettingsAsync(string voiceId, VoiceSettings voiceSettings, CancellationToken cancellationToken = default)
         {
+            if (string.IsNullOrWhiteSpace(voiceId))
+            {
+                throw new ArgumentNullException(nameof(voiceId));
+            }
+
             var payload = JsonConvert.SerializeObject(voiceSettings).ToJsonStringContent();
             var response = await Api.Client.PostAsync(GetUrl($"/{voiceId}/settings/edit"), payload, cancellationToken);
             await response.ReadAsStringAsync();
@@ -240,6 +255,11 @@ namespace ElevenLabs.Voices
         /// <returns>True, if voice was successfully deleted.</returns>
         public async Task<bool> DeleteVoiceAsync(string voiceId, CancellationToken cancellationToken = default)
         {
+            if (string.IsNullOrWhiteSpace(voiceId))
+            {
+                throw new ArgumentNullException(nameof(voiceId));
+            }
+
             var response = await Api.Client.DeleteAsync(GetUrl($"/{voiceId}"), cancellationToken);
             await response.CheckResponseAsync();
             return response.IsSuccessStatusCode;
@@ -257,6 +277,16 @@ namespace ElevenLabs.Voices
         /// <returns><see cref="AudioClip"/>.</returns>
         public async Task<AudioClip> GetVoiceSampleAsync(string voiceId, string sampleId, string saveDirectory = null, CancellationToken cancellationToken = default)
         {
+            if (string.IsNullOrWhiteSpace(voiceId))
+            {
+                throw new ArgumentNullException(nameof(voiceId));
+            }
+
+            if (string.IsNullOrWhiteSpace(sampleId))
+            {
+                throw new ArgumentNullException(nameof(sampleId));
+            }
+
             var response = await Api.Client.GetAsync(GetUrl($"/{voiceId}/samples/{sampleId}/audio"), cancellationToken);
             await response.CheckResponseAsync();
 
@@ -314,6 +344,16 @@ namespace ElevenLabs.Voices
         /// <returns>True, if <see cref="Voice"/> <see cref="Sample"/> was successfully deleted.</returns>
         public async Task<bool> DeleteVoiceSampleAsync(string voiceId, string sampleId, CancellationToken cancellationToken = default)
         {
+            if (string.IsNullOrWhiteSpace(voiceId))
+            {
+                throw new ArgumentNullException(nameof(voiceId));
+            }
+
+            if (string.IsNullOrWhiteSpace(sampleId))
+            {
+                throw new ArgumentNullException(nameof(sampleId));
+            }
+
             var response = await Api.Client.DeleteAsync(GetUrl($"/{voiceId}/samples/{sampleId}"), cancellationToken);
             await response.CheckResponseAsync();
             return response.IsSuccessStatusCode;
